LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/scrollIntoView-horizontal-partially-visible-expected.txt


FAIL scrollIntoView scrolls partially-visible child in both axes assert_equals: Should have scrolled in the inline direction expected 200 but got 0
